+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageHTMLJSInterface.h
@@ -1,7 +1,0 @@
-#import <Foundation/Foundation.h>
-
-@interface ABKInAppMessageHTMLJSInterface : NSObject
-
-+ (NSString *)getJSInterface;
-
-@end