+++ AppboyUI/ABKUIUtils/ABKUIUtils.h
@@ -32,4 +32,12 @@
                              darkColor:(UIColor *)darkColor;
 + (BOOL)string:(NSString *)string1 isEqualToString:(NSString *)string2;
 
+/*!
+ * Verifies that one of the responders in the responder chain is kind of class aClass.
+ * @param responder The start of the UIResponder chain
+ * @param aClass The UIResponder subclass looked for in the responder chain
+ * @return YES if aClass is found in the responder chain, NO otherwise
+ */
++ (BOOL)responderChainOf:(UIResponder *)responder hasKindOfClass:(Class)aClass;
+
 @end