+++ Example/Stopwatch/Sources/Views/StopwatchButton.m
@@ -14,4 +14,10 @@
   return self;
 }
 
+#pragma mark - NSSecureCoding
+
++ (BOOL)supportsSecureCoding {
+  return YES;
+}
+
 @end