+++ Example/Stopwatch/Sources/Views/StopwatchSegmentedControl.m
@@ -11,4 +11,10 @@
   return self;
 }
 
+#pragma mark - NSSecureCoding
+
++ (BOOL)supportsSecureCoding {
+  return YES;
+}
+
 @end