+++ Samples/ContentCards/BrazeContentCardsSampleApp/BrazeContentCardsSampleApp/ViewController.m
@@ -2,6 +2,10 @@
 #import "AppboyContentCards.h"
 #import "CustomContentCardsTableViewController.h"
 
+@interface ViewController () <ABKContentCardsTableViewControllerDelegate>
+
+@end
+
 @implementation ViewController
 
 - (void)viewDidLoad {
@@ -14,6 +18,7 @@
 
 - (IBAction)displayModalContentCards:(id)sender {
   ABKContentCardsViewController *contentCards = [[ABKContentCardsViewController alloc] init];
+  contentCards.contentCardsViewController.delegate = self;
   contentCards.contentCardsViewController.disableUnreadIndicator = !self.unreadIndicatorSwitch.on;
   contentCards.contentCardsViewController.navigationItem.title = @"Modal Cards";
   [self presentViewController:contentCards animated:YES completion:nil];
@@ -21,6 +26,7 @@
 
 - (IBAction)displayNavigationContentCards:(id)sender {
   ABKContentCardsTableViewController *contentCards = [ABKContentCardsTableViewController getNavigationContentCardsViewController];
+  contentCards.delegate = self;
   contentCards.disableUnreadIndicator = !self.unreadIndicatorSwitch.on;
   contentCards.navigationItem.title = @"Navigation Cards";
   [self.navigationController pushViewController:contentCards animated:YES];
@@ -28,14 +34,29 @@
 
 - (IBAction)displayCustomContentCards:(id)sender {
   CustomContentCardsTableViewController *customContentCards = [[CustomContentCardsTableViewController alloc] init];
+  customContentCards.delegate = self;
   customContentCards.disableUnreadIndicator = !self.unreadIndicatorSwitch.on;
   [self.navigationController pushViewController:customContentCards animated:YES];
 }
+
 - (IBAction)changeUser:(id)sender {
   if (self.userTextField.text.length > 0) {
     NSString *userID = self.userTextField.text;
     [[Appboy sharedInstance] changeUser:userID];
   }
+}
+
+#pragma mark - ABKContentCardsTableViewControllerDelegate
+
+- (BOOL)contentCardTableViewController:(ABKContentCardsTableViewController *)viewController
+                 shouldHandleCardClick:(NSURL *)url {
+  NSLog(@"%@", NSStringFromSelector(_cmd));
+  return YES;
+}
+
+- (void)contentCardTableViewController:(ABKContentCardsTableViewController *)viewController
+                    didHandleCardClick:(NSURL *)url {
+  NSLog(@"%@", NSStringFromSelector(_cmd));
 }
 
 @end